08_Predictive_Diagnostics/requirements.txt
streamlit==1.28.0
pandas==2.1.1
numpy==1.24.3
xgboost==2.0.0
scikit-learn==1.3.1
plotly==5.17.0
openpyxl==3.1.2
joblib==1.3.2
python-dotenv==1.0.0
googlemaps==4.10.0
folium==0.15.1
streamlit-folium==0.18.0
openai>=1.0.0